+++ src/Checkbox.js
@@ -9,6 +9,7 @@ import { bsClass, getClassSet, prefix, splitBsProps }
 const propTypes = {
   inline: PropTypes.bool,
   disabled: PropTypes.bool,
+  title: PropTypes.string,
   /**
    * Only valid if `inline` is not set.
    */
@@ -28,6 +29,7 @@ const propTypes = {
 const defaultProps = {
   inline: false,
   disabled: false,
+  title: '',
 };
 
 class Checkbox extends React.Component {
@@ -39,6 +41,7 @@ class Checkbox extends React.Component {
       inputRef,
       className,
       style,
+      title,
       children,
       ...props
     } = this.props;
@@ -70,7 +73,7 @@ class Checkbox extends React.Component {
       );
 
       return (
-        <label className={classNames(className, classes)} style={style}>
+        <label className={classNames(className, classes)} style={style} title={title}>
           {input}
           {children}
         </label>
@@ -87,7 +90,7 @@ class Checkbox extends React.Component {
 
     return (
       <div className={classNames(className, classes)} style={style}>
-        <label>
+        <label title={title}>
           {input}
           {children}
         </label>