+++ src/Thumbnail.js
@@ -6,14 +6,31 @@ import SafeAnchor from './SafeAnchor';
 import { bsClass, getClassSet, splitBsProps } from './utils/bootstrapUtils';
 
 const propTypes = {
+  /**
+   * src property that is passed down to the image inside this component
+   */
   src: PropTypes.string,
+  /**
+   * alt property that is passed down to the image inside this component
+   */
   alt: PropTypes.string,
+  /**
+   * href property that is passed down to the image inside this component
+   */
   href: PropTypes.string,
+  /**
+   * onError callback that is passed down to the image inside this component
+   */
+  onError: PropTypes.func,
+  /**
+   * onLoad callback that is passed down to the image inside this component
+   */
+  onLoad: PropTypes.func
 };
 
 class Thumbnail extends React.Component {
   render() {
-    const { src, alt, className, children, ...props } = this.props;
+    const { src, alt, onError, onLoad, className, children, ...props } = this.props;
     const [bsProps, elementProps] = splitBsProps(props);
 
     const Component = elementProps.href ? SafeAnchor : 'div';
@@ -24,7 +41,7 @@ class Thumbnail extends React.Component {
         {...elementProps}
         className={classNames(className, classes)}
       >
-        <img src={src} alt={alt} />
+        <img {...{src, alt, onError, onLoad}} />
 
         {children && (
           <div className="caption">