+++ test/DropdownSpec.js
@@ -1,7 +1,7 @@
 import keycode from 'keycode';
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
 import tsp from 'teaspoon';
 
 import Dropdown from '../src/Dropdown';