+++ test/ImageSpec.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
+
 import Image from '../src/Image';
 
 describe('Image', () => {