+++ test/ThumbnailSpec.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
 
 import Thumbnail from '../src/Thumbnail';
 
@@ -58,4 +58,25 @@ describe('<Thumbnail>', () => {
     );
     assert.ok(ReactDOM.findDOMNode(instance).lastChild.className.match(/\bcaption\b/));
   });
+
+  it('Should have an img with an onError callback', () => {
+    const onErrorSpy = sinon.spy();
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Thumbnail href="#" src="#" alt="test" onError={onErrorSpy} />
+    );
+    const img = ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'img');
+    ReactTestUtils.Simulate.error(img);
+    expect(onErrorSpy).to.have.been.calledOnce;
+  });
+
+  it('Should have an img with an onLoad callback', () => {
+    const onLoadSpy = sinon.spy();
+    const instance = ReactTestUtils.renderIntoDocument(
+      <Thumbnail href="#" src="#" alt="test" onLoad={onLoadSpy} />
+    );
+    const img = ReactTestUtils.findRenderedDOMComponentWithTag(instance, 'img');
+    ReactTestUtils.Simulate.load(img);
+    expect(onLoadSpy).to.have.been.calledOnce;
+  });
+
 });