+++ test/index.js
@@ -4,7 +4,7 @@ import { _resetWarned } from '../src/utils/deprecationWarning';
 
 beforeEach(() => {
   /* eslint-disable no-console */
-  sinon.stub(console, 'error', msg => {
+  sinon.stub(console, 'error').callsFake((msg) => {
     let expected = false;
 
     console.error.expected.forEach(about => {